+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from 'next';
 import { Geist, Geist_Mono } from 'next/font/google';
 import './globals.css';
 import { AuthProvider } from "@/components/providers/AuthProvider";
-import Header from '@/components/Header';
+import Header from '@/components/header/Header';
 
 const geistSans = Geist({
   variable: '--font-geist-sans',
